requirements.txt
scikit-learn
pandas
numpy
spacy
nltk
sentence-transformers
PyPDF2
python-docx
streamlit